+++ NINA/View/Equipment/FlatDevice/PegasusAstroFlatMasterSetupView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.IO.Ports;
+using System.Windows.Controls;
+
+namespace NINA.View.Equipment {
+
+    /// <summary>
+    /// Interaction logic for PegasusAstroFlatMasterSetupView.xaml
+    /// </summary>
+    public partial class PegasusAstroFlatMasterSetupView : UserControl {
+
+        public PegasusAstroFlatMasterSetupView() {
+            InitializeComponent();
+        }
+    }
+}