+++ RELEASE_NOTES.md
@@ -24,6 +24,8 @@
    - set the voltage of the variable power port
    - control the dew heater ports, including setting the Auto-Dew feature
    - support for using the powerbox sensors as a weather device
+ - **Pegasus Astro Flatmaster Support**
+   - connect and control the Pegasus Astro Flatmaster from within the application
 
 
 ## Localization